backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==3.2.2",
    "fastapi>=0.133.0",
    "langchain-google-genai>=4.2.1",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pymongo>=4.16.0",
    "pymupdf>=1.27.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.41.0",
]
